refactor: convert Core Data model to SwiftData model; combine all code related to Pokemon model
