+++ Dex/Pokemon.swift
@@ -1,24 +1,33 @@
 //
-//  FetchedPokemon.swift
+//  Pokemon.swift
 //  Dex
 //
-//  Created by Elchin Ismailov on 24.03.2025.
+//  Created by Ismailov on 4.04.2025.
+//
 //
 
 import Foundation
+import SwiftData
+import SwiftUI
 
-struct FetchedPokemon: Decodable {
-    let id: Int16
-    let name: String
-    let types: [String]
-    let hp: Int16
-    let attack: Int16
-    let defense: Int16
-    let specialAttack: Int16
-    let specialDefense: Int16
-    let speed: Int16
-    let spriteURL: URL
-    let shinyURL: URL
+
+@Model
+class Pokemon: Decodable {
+    
+    @Attribute(.unique) var id: Int
+    var name: String
+    var types: [String]
+    var hp: Int
+    var attack: Int
+    var defense: Int
+    var specialAttack: Int
+    var specialDefense: Int
+    var speed: Int
+    var spriteURL: URL
+    var shinyURL: URL
+    var sprite: Data?
+    var shiny: Data?
+    var favorite: Bool = false
     
     enum CodingKeys: CodingKey {
         case id
@@ -45,10 +54,10 @@ struct FetchedPokemon: Decodable {
         }
     }
     
-    init(from decoder: any Decoder) throws {
+    required init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         
-        self.id = try container.decode(Int16.self, forKey: .id)
+        self.id = try container.decode(Int.self, forKey: .id)
         self.name = try container.decode(String.self, forKey: .name)
         
         var decodedTypes: [String] = []
@@ -66,11 +75,11 @@ struct FetchedPokemon: Decodable {
         
         self.types = decodedTypes
         
-        var decodedStats: [Int16] = []
+        var decodedStats: [Int] = []
         var statsContainer = try container.nestedUnkeyedContainer(forKey: .stats)
         while !statsContainer.isAtEnd {
             let statsDictionaryContainer = try statsContainer.nestedContainer(keyedBy: CodingKeys.StatDictionaryKeys.self)
-            let stat = try statsDictionaryContainer.decode(Int16.self, forKey: .baseStat)
+            let stat = try statsDictionaryContainer.decode(Int.self, forKey: .baseStat)
             decodedStats.append(stat)
         }
         self.hp = decodedStats[0]
@@ -83,5 +92,63 @@ struct FetchedPokemon: Decodable {
         let spriteContainer = try container.nestedContainer(keyedBy: CodingKeys.SpriteKeys.self, forKey: .sprites)
         self.spriteURL = try spriteContainer.decode(URL.self, forKey: .spriteURL)
         self.shinyURL = try spriteContainer.decode(URL.self, forKey: .shinyURL)
+    }
+    
+    var spriteImage: Image {
+        if let data = sprite, let image = UIImage(data: data) {
+            Image(uiImage: image)
+        } else {
+            Image(.bulbasaur)
+        }
+    }
+    
+    var shinyImage: Image {
+        if let data = shiny, let image = UIImage(data: data) {
+            Image(uiImage: image)
+        } else {
+            Image(.shinybulbasaur)
+        }
+    }
+    
+    var background: ImageResource {
+        switch types[0] {
+        case "rock", "ground", "steel", "fighting", "ghost", "dark", "psychic":
+                .rockgroundsteelfightingghostdarkpsychic
+        case "fire", "dragon":
+                .firedragon
+        case "flying", "bug":
+                .flyingbug
+        case "ice":
+                .ice
+        case "water":
+                .water
+        default:
+                .normalgrasselectricpoisonfairy
+        }
+    }
+    
+    var typeColor: Color {
+        Color(types[0].capitalized)
+    }
+    
+    var stats: [Stat] {
+        [
+            Stat(id: 1, name: "HP", value: hp),
+            Stat(id: 2, name: "Attack", value: attack),
+            Stat(id: 3, name: "Defense", value: defense),
+            Stat(id: 4, name: "Special Attack", value: specialAttack),
+            Stat(id: 5, name: "Special Defense", value: specialDefense),
+            Stat(id: 6, name: "Speed", value: speed)
+        ]
+    }
+    
+    var highestStat: Stat {
+        stats.max { $0.value < $1.value }!
+    }
+    
+    struct Stat: Identifiable {
+        let id: Int
+        let name: String
+        let value: Int
     }
 }
